Build tweaks to remove cchost
